+++ src/utils/v2v3/loadV2V3Contract.ts
@@ -4,17 +4,12 @@ import { NetworkName } from 'models/networkName'
 import { SignerOrProvider } from 'models/signerOrProvider'
 import { V2V3ContractName } from 'models/v2v3/contracts'
 import { CV2V3 } from 'models/v2v3/cv'
+import { loadJBTiered721DelegateProjectDeployerContract } from '../../hooks/JB721Delegate/contracts/JBTiered721DelegateProjectDeployer'
 import { loadJBProjectHandlesContract } from './contractLoaders/JBProjectHandles'
-import { loadJBTiered721DelegateProjectDeployerContract } from './contractLoaders/JBTiered721DelegateProjectDeployer'
 import { loadJuiceboxV2OrV3Contract } from './contractLoaders/JuiceboxV2OrV3'
 import { loadPublicResolverContract } from './contractLoaders/PublicResolver'
 import { loadVeNftDeployer } from './contractLoaders/VeNftDeployer'
 import { loadVeTokenUriResolver } from './contractLoaders/VeTokenUriResolver'
-
-export interface ForgeDeploy {
-  receipts: { contractAddress: string }[]
-  transactions: { contractName: string; contractAddress: string }[]
-}
 
 export const loadV2V3Contract = async (
   contractName: V2V3ContractName,
